Change cfl_method to be "no"
The cfl_method default is "cfl". Here, I change it so that the DamBreak scenario used "no" clf_method.

A reminder: "cfl_method" is a parameter that allows SlishSplash to use adaptive time-step or no adaptive time-step.

This is important for Maria dataset creation, but I wanted to check how the simulations are visualised in this sense with different types of fluid.