Add low-level open foam
In this PR, I add the low-level call to openfoam.

I would like to notice that I am introducing the run_simulation in `inductiva/openfoam.py` since there seems to be no appropriate place for it as of now. Openfoam is a very general simulator, so maybe we can a folder for computational fluid dynamics as `cfd` and include openfoam.py inside as well as all the others concerning fluids, sw, sph.