Add visualization capabilities to the WindTunnel scenario
This commit starts implementing the visualizations of the outputs provided by the WindTunnel scenario. Recall, that OpenFOAM has the ability to post-process the information and output some this results in various forms.

With this in mind, we use pyvista to implement some of the visualization capabilities namely: scalar fields over the object, streamlines and cutting plane.