Simplify implementation of WindTunnel Outputs
This refactors the implementation of the WindTunnelSImulationOutput with the Sarmento comments in mind. Moreover, it already adds the streamlines and cutting planes visuals. It still needs a bit of clean-up and work, but I think the flow is what we want right now!

To plot the pressure field, one does:
```
pressure_field = sim_output.get_pressure_field(time_step)

pressure_field.render()
```