feat: automatically generate sitemap.xml
These changes will ensure that when the static documentation
files are generated a 'sitemap.xml' file is also built.

This is done using sphinx-sitemap (a Sphinx plugin)
that generates sitemaps.org compliant sitemaps.

Quoting sitemaps.org:

"Sitemaps are an easy way for webmasters to inform
search engines about pages on their sites that are
available for crawling."

A helper function was created in order to determine
the correct URL scheme. It translates the current
git branch name to a version string:

* 'latest' if branch is main
* {branch_name} if branch is NOT main
* 'unknown_version' if an error occurs

See https://sphinx-sitemap.readthedocs.io/en/latest/advanced-configuration.html
for details